+++ sdk/resourcemanager/mediaservices/armmediaservices/zz_generated_liveoutputs_client.go
@@ -54,6 +54,68 @@ func NewLiveOutputsClient(subscriptionID string, credential azcore.TokenCredenti
 	return client, nil
 }
 
+// AsyncOperation - Get a Live Output operation status.
+// If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-11-01
+// resourceGroupName - The name of the resource group within the Azure subscription.
+// accountName - The Media Services account name.
+// operationID - The ID of an ongoing async operation.
+// options - LiveOutputsClientAsyncOperationOptions contains the optional parameters for the LiveOutputsClient.AsyncOperation
+// method.
+func (client *LiveOutputsClient) AsyncOperation(ctx context.Context, resourceGroupName string, accountName string, operationID string, options *LiveOutputsClientAsyncOperationOptions) (LiveOutputsClientAsyncOperationResponse, error) {
+	req, err := client.asyncOperationCreateRequest(ctx, resourceGroupName, accountName, operationID, options)
+	if err != nil {
+		return LiveOutputsClientAsyncOperationResponse{}, err
+	}
+	resp, err := client.pl.Do(req)
+	if err != nil {
+		return LiveOutputsClientAsyncOperationResponse{}, err
+	}
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		return LiveOutputsClientAsyncOperationResponse{}, runtime.NewResponseError(resp)
+	}
+	return client.asyncOperationHandleResponse(resp)
+}
+
+// asyncOperationCreateRequest creates the AsyncOperation request.
+func (client *LiveOutputsClient) asyncOperationCreateRequest(ctx context.Context, resourceGroupName string, accountName string, operationID string, options *LiveOutputsClientAsyncOperationOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Media/mediaservices/{accountName}/liveOutputOperations/{operationId}"
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	if resourceGroupName == "" {
+		return nil, errors.New("parameter resourceGroupName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{resourceGroupName}", url.PathEscape(resourceGroupName))
+	if accountName == "" {
+		return nil, errors.New("parameter accountName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{accountName}", url.PathEscape(accountName))
+	if operationID == "" {
+		return nil, errors.New("parameter operationID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{operationId}", url.PathEscape(operationID))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.host, urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2021-11-01")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
+// asyncOperationHandleResponse handles the AsyncOperation response.
+func (client *LiveOutputsClient) asyncOperationHandleResponse(resp *http.Response) (LiveOutputsClientAsyncOperationResponse, error) {
+	result := LiveOutputsClientAsyncOperationResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.AsyncOperationResult); err != nil {
+		return LiveOutputsClientAsyncOperationResponse{}, err
+	}
+	return result, nil
+}
+
 // BeginCreate - Creates a new live output.
 // If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2021-11-01
@@ -335,6 +397,78 @@ func (client *LiveOutputsClient) listHandleResponse(resp *http.Response) (LiveOu
 	result := LiveOutputsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.LiveOutputListResult); err != nil {
 		return LiveOutputsClientListResponse{}, err
+	}
+	return result, nil
+}
+
+// OperationLocation - Get a Live Output operation status.
+// If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-11-01
+// resourceGroupName - The name of the resource group within the Azure subscription.
+// accountName - The Media Services account name.
+// liveEventName - The name of the live event, maximum length is 32.
+// liveOutputName - The name of the live output.
+// operationID - The ID of an ongoing async operation.
+// options - LiveOutputsClientOperationLocationOptions contains the optional parameters for the LiveOutputsClient.OperationLocation
+// method.
+func (client *LiveOutputsClient) OperationLocation(ctx context.Context, resourceGroupName string, accountName string, liveEventName string, liveOutputName string, operationID string, options *LiveOutputsClientOperationLocationOptions) (LiveOutputsClientOperationLocationResponse, error) {
+	req, err := client.operationLocationCreateRequest(ctx, resourceGroupName, accountName, liveEventName, liveOutputName, operationID, options)
+	if err != nil {
+		return LiveOutputsClientOperationLocationResponse{}, err
+	}
+	resp, err := client.pl.Do(req)
+	if err != nil {
+		return LiveOutputsClientOperationLocationResponse{}, err
+	}
+	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusAccepted) {
+		return LiveOutputsClientOperationLocationResponse{}, runtime.NewResponseError(resp)
+	}
+	return client.operationLocationHandleResponse(resp)
+}
+
+// operationLocationCreateRequest creates the OperationLocation request.
+func (client *LiveOutputsClient) operationLocationCreateRequest(ctx context.Context, resourceGroupName string, accountName string, liveEventName string, liveOutputName string, operationID string, options *LiveOutputsClientOperationLocationOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Media/mediaservices/{accountName}/liveEvents/{liveEventName}/liveOutputs/{liveOutputName}/operationLocations/{operationId}"
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	if resourceGroupName == "" {
+		return nil, errors.New("parameter resourceGroupName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{resourceGroupName}", url.PathEscape(resourceGroupName))
+	if accountName == "" {
+		return nil, errors.New("parameter accountName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{accountName}", url.PathEscape(accountName))
+	if liveEventName == "" {
+		return nil, errors.New("parameter liveEventName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{liveEventName}", url.PathEscape(liveEventName))
+	if liveOutputName == "" {
+		return nil, errors.New("parameter liveOutputName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{liveOutputName}", url.PathEscape(liveOutputName))
+	if operationID == "" {
+		return nil, errors.New("parameter operationID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{operationId}", url.PathEscape(operationID))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.host, urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2021-11-01")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
+// operationLocationHandleResponse handles the OperationLocation response.
+func (client *LiveOutputsClient) operationLocationHandleResponse(resp *http.Response) (LiveOutputsClientOperationLocationResponse, error) {
+	result := LiveOutputsClientOperationLocationResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.LiveOutput); err != nil {
+		return LiveOutputsClientOperationLocationResponse{}, err
 	}
 	return result, nil
 }