+++ sdk/resourcemanager/mediaservices/armmediaservices/zz_generated_models.go
@@ -449,6 +449,30 @@ type AssetsClientUpdateOptions struct {
 	// placeholder for future optional parameters
 }
 
+// AsyncOperationErrorDetail - The error detail of an async operation result.
+type AsyncOperationErrorDetail struct {
+	// The error code.
+	Code *string `json:"code,omitempty"`
+
+	// The error message.
+	Message *string `json:"message,omitempty"`
+
+	// Id of the entity on which the operation is performed.
+	Target *string `json:"target,omitempty"`
+}
+
+// AsyncOperationResult - The status of an async operation.
+type AsyncOperationResult struct {
+	// The error object
+	Error *AsyncOperationErrorDetail `json:"error,omitempty"`
+
+	// Operation Id of the async operation.
+	Name *string `json:"name,omitempty"`
+
+	// Operation status of the async operation.
+	Status *AsyncOperationStatus `json:"status,omitempty"`
+}
+
 // AudioClassification provides polymorphic access to related types.
 // Call the interface's GetAudio() method to access the common type.
 // Use a type switch to determine the concrete type.  The possible types are:
@@ -2953,6 +2977,11 @@ type LiveEventTranscription struct {
 	OutputTranscriptionTrack *LiveEventOutputTranscriptionTrack `json:"outputTranscriptionTrack,omitempty"`
 }
 
+// LiveEventsClientAsyncOperationOptions contains the optional parameters for the LiveEventsClient.AsyncOperation method.
+type LiveEventsClientAsyncOperationOptions struct {
+	// placeholder for future optional parameters
+}
+
 // LiveEventsClientBeginAllocateOptions contains the optional parameters for the LiveEventsClient.BeginAllocate method.
 type LiveEventsClientBeginAllocateOptions struct {
 	// Resumes the LRO from the provided token.
@@ -3004,6 +3033,11 @@ type LiveEventsClientGetOptions struct {
 
 // LiveEventsClientListOptions contains the optional parameters for the LiveEventsClient.List method.
 type LiveEventsClientListOptions struct {
+	// placeholder for future optional parameters
+}
+
+// LiveEventsClientOperationLocationOptions contains the optional parameters for the LiveEventsClient.OperationLocation method.
+type LiveEventsClientOperationLocationOptions struct {
 	// placeholder for future optional parameters
 }
 
@@ -3072,6 +3106,11 @@ type LiveOutputProperties struct {
 	ResourceState *LiveOutputResourceState `json:"resourceState,omitempty" azure:"ro"`
 }
 
+// LiveOutputsClientAsyncOperationOptions contains the optional parameters for the LiveOutputsClient.AsyncOperation method.
+type LiveOutputsClientAsyncOperationOptions struct {
+	// placeholder for future optional parameters
+}
+
 // LiveOutputsClientBeginCreateOptions contains the optional parameters for the LiveOutputsClient.BeginCreate method.
 type LiveOutputsClientBeginCreateOptions struct {
 	// Resumes the LRO from the provided token.
@@ -3091,6 +3130,12 @@ type LiveOutputsClientGetOptions struct {
 
 // LiveOutputsClientListOptions contains the optional parameters for the LiveOutputsClient.List method.
 type LiveOutputsClientListOptions struct {
+	// placeholder for future optional parameters
+}
+
+// LiveOutputsClientOperationLocationOptions contains the optional parameters for the LiveOutputsClient.OperationLocation
+// method.
+type LiveOutputsClientOperationLocationOptions struct {
 	// placeholder for future optional parameters
 }
 
@@ -4117,6 +4162,12 @@ type StreamingEndpointSKUInfoListResult struct {
 	Value []*ArmStreamingEndpointSKUInfo `json:"value,omitempty"`
 }
 
+// StreamingEndpointsClientAsyncOperationOptions contains the optional parameters for the StreamingEndpointsClient.AsyncOperation
+// method.
+type StreamingEndpointsClientAsyncOperationOptions struct {
+	// placeholder for future optional parameters
+}
+
 // StreamingEndpointsClientBeginCreateOptions contains the optional parameters for the StreamingEndpointsClient.BeginCreate
 // method.
 type StreamingEndpointsClientBeginCreateOptions struct {
@@ -4167,6 +4218,12 @@ type StreamingEndpointsClientGetOptions struct {
 
 // StreamingEndpointsClientListOptions contains the optional parameters for the StreamingEndpointsClient.List method.
 type StreamingEndpointsClientListOptions struct {
+	// placeholder for future optional parameters
+}
+
+// StreamingEndpointsClientOperationLocationOptions contains the optional parameters for the StreamingEndpointsClient.OperationLocation
+// method.
+type StreamingEndpointsClientOperationLocationOptions struct {
 	// placeholder for future optional parameters
 }